requirements.txt
aiohttp==3.6.2
alembic==1.0.8
appnope==0.1.0
asgiref==3.2.7
async-timeout==3.0.1
atomicwrites==1.3.0
attrdict==2.0.1
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==3.1.0
certifi==2019.3.9
chardet==3.0.4
click==7.1.1
cycler==0.10.0
dash==1.10.0
dash-core-components==1.9.0
dash-daq==0.4.0
dash-html-components==1.0.3
dash-renderer==1.3.0
dash-table==4.6.2
dash-table-experiments==0.6.0
decorator==4.3.2
defusedxml==0.5.0
Django==3.0.5
django-bootstrap4==1.1.1
django-plotly-dash==1.1.5
dpd-components==0.1.0
dpd-static-support==0.0.5
entrypoints==0.3
filterpy==1.4.1
Flask==1.1.1
Flask-Compress==1.4.0
future==0.18.2
idna==2.8
imageio==2.6.1
ipykernel==5.1.3
ipython==7.10.2
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-book==0.4.2
jupyter-client==5.3.4
jupyter-console==5.2.0
jupyter-core==4.6.1
jupyter-plotly-dash==0.4.2
kiwisolver==1.0.1
llvmlite==0.30.0
Mako==1.0.7
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
more-itertools==6.0.0
multidict==4.7.5
nbclean==0.3.2
nbconvert==5.4.1
nbformat==4.4.0
nbgrader==0.5.5
nbserverproxy==0.8.8
networkx==2.4
notebook==5.7.6
numba==0.46.0
numpy==1.16.2
opencv-python==4.0.0.21
pandas==1.1.4
pandocfilters==1.4.2
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
plotly==4.6.0
pluggy==0.9.0
prometheus-client==0.6.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
py==1.8.0
Pygments==2.5.2
pyparsing==2.3.1
pyrsistent==0.14.11
pytest==4.3.1
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2018.9
PyWavelets==1.1.1
PyYAML==5.1
pyzmq==18.1.1
qtconsole==4.4.3
requests==2.21.0
retrying==1.3.3
ruamel.yaml==0.15.89
scikit-image==0.16.2
scikit-learn==0.22
scipy==1.4.0
seaborn==0.11.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.12.0
soupsieve==2.0
SQLAlchemy==1.3.1
sqlparse==0.3.1
terminado==0.8.1
testpath==0.4.2
tornado==6.0.3
tqdm==4.31.1
traitlets==4.3.3
urllib3==1.24.1
virtualenv==16.4.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.4.2
xarray==0.16.1
yarl==1.4.2
